Make dataset entry tools editable
